pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "redsl"
version = "1.2.45"
description = "ReDSL — Refactor + DSL + Self-Learning. LLM-powered autonomous code refactoring."
readme = "README_EN.md"
requires-python = ">=3.11"
license = "Apache-2.0"
keywords = ["refactoring", "llm", "dsl", "code-quality", "ai"]
authors = [
    {name = "ReDSL Team", email = "redsl@example.com"},
    {name = "Tom Sapletta", email = "tom@sapletta.com"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.44.0",
    "pydantic>=2.10.0",
    "litellm>=1.52.0",
    "chromadb>=0.6.0",
    "pyyaml>=6.0.2",
    "rich>=13.9.0",
    "httpx>=0.28.0",
    "click>=8.1.7",
    "python-dotenv>=1.0.1",
    "goal>=2.1.0",
    "costs>=0.1.20",
    "pfix>=0.1.60",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0",
    "pytest-xdist>=3.5",
    "pytest-cov>=7.1",
    "ruff>=0.15.0",
    "mypy>=1.13",
    "goal>=2.1.0",
    "costs>=0.1.20",
    "pfix>=0.1.60",
]

[project.scripts]
redsl = "redsl.cli.__init__:cli"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "--tb=short -n auto --no-header -m 'not slow and not e2e'"
markers = [
    "slow: marks tests as slow (>5s)",
    "integration: integration tests requiring external tools",
    "e2e: end-to-end tests running full workflows without mocks",
]

[tool.vallm]
exclude = ["tests/", "archive/", "*_test.py", "test_*.py", "conftest.py"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true

[tool.setuptools.packages.find]
exclude = ["config", "tests"]

[tool.setuptools.package-data]
# YAMLs live in examples/ directory, not inside the package

[tool.pfix]
# Self-healing Python configuration
model = "openrouter/openai/gpt-5-mini"
auto_apply = true
auto_install_deps = true
auto_restart = false
max_retries = 3
create_backups = false
git_auto_commit = false

[tool.pfix.runtime_todo]
enabled = true
todo_file = "TODO.md"
min_severity = "low"
deduplicate = true

[tool.costs]
# AI Cost tracking configuration
badge = true
update_readme = true
readme_path = "README.md"
default_model = "openrouter/openai/gpt-5-mini"
analysis_mode = "byok"
full_history = true
max_commits = 500

# Cost thresholds for badge colors (USD)
badge_color_thresholds = { low = 1.0, medium = 5.0, high = 10.0, critical = 50.0 }
